# _reversa_sdd/openapi/zenith-api.yaml
openapi: 3.1.0
info:
  title: Zenith AI Audio Hub API
  version: 1.0.0
  description: API REST para gerenciamento de PBXs e WebSocket para agent assist

servers:
  - url: https://zenith.exemplo.com
    description: Produção
  - url: http://localhost:8000
    description: Desenvolvimento

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /ready:
    get:
      summary: Readiness check
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready

  /api/v1/admin/pbxs:
    post:
      summary: Criar PBX
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PBXCreate"
      responses:
        "201":
          description: PBX criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PBXResponse"
        "400":
          description: Dados inválidos
        "403":
          description: Apenas tenant_admin
        "409":
          description: PBX já existe

    get:
      summary: Listar PBXs do tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de PBXs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PBXResponse"
        "403":
          description: Apenas tenant_admin

  /ws/{call_id}:
    get:
      summary: WebSocket Agent Assist
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    PBXCreate:
      type: object
      required:
        - name
        - host
      properties:
        name:
          type: string
          maxLength: 128
        host:
          type: string
          maxLength: 128
        port:
          type: integer
          minimum: 1
          maximum: 65535
          default: 5060

    PBXResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        created_at:
          type: string
          format: date-time